+++ src/Language/PHP/Parser/Expression.hs
@@ -256,7 +256,14 @@ parseExprWithContextAndBody parseBody pMember = parseExprRec
       , (void (lexeme (M.try (C.char '>' <* M.notFollowedBy (C.char '>' <|> C.char '=')))), OpGt)
       ]
 
-    parsePipe = parseBinaryLeft parseShift [ (void (symbol "|>"), OpPipe) ]
+    parsePipe = parseBinaryLeft parseConcat [ (void (symbol "|>"), OpPipe) ]
+
+    -- Since PHP 8.0, "." sits on its own precedence level, below "<<"/">>"
+    -- and above comparison operators (RFC: Change the precedence of the
+    -- concatenation operator), so "a" . 1 + 2 parses as "a" . (1 + 2).
+    parseConcat = parseBinaryLeft parseShift
+      [ (void (lexeme (M.try (C.char '.' <* M.notFollowedBy (C.char '.' <|> C.char '=')))), OpConcat)
+      ]
 
     parseShift = parseBinaryLeft parseAddSub
       [ (void (symbol "<<"), OpShiftLeft)
@@ -266,7 +273,6 @@ parseExprWithContextAndBody parseBody pMember = parseExprRec
     parseAddSub = parseBinaryLeft parseMulDivMod
       [ (void (lexeme (M.try (C.char '+' <* M.notFollowedBy (C.char '+' <|> C.char '=')))), OpAdd)
       , (void (lexeme (M.try (C.char '-' <* M.notFollowedBy (C.char '-' <|> C.char '>' <|> C.char '=')))), OpSub)
-      , (void (lexeme (M.try (C.char '.' <* M.notFollowedBy (C.char '.' <|> C.char '=')))), OpConcat)
       ]
 
     -- @instanceof@ binds below unary operators and exponentiation, but above